Files are structured
 - screens, localdata, servies are three different folder to contain respective files
 - screens contains UI elements
 -services contains the logic part
 - localdata contains the files for state management (future use)
